AG/requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
IPython>=8.0.0 